+++ index.js
@@ -1,3 +1,5 @@
+const getVerifiedPaymentDetails = require("./lib/verify");
 const getQRCodeUrl = require("./lib/qrCode");
 
+module.exports = getVerifiedPaymentDetails;
 module.exports = getQRCodeUrl;
