+++ lib/apiRequest.js
@@ -1,0 +1,42 @@
+const fetch = require("node-fetch");
+const getTime = require('./time');
+const generateSignature = require('./generateSignature');
+const CONSTANTS = require("../utils/constants");
+  
+async function apiRequest(apiSecret, apiKey, method, path, body){  
+  const url = `${CONSTANTS.baseUrl}${path}`
+  const path = `${CONSTANTS.apiPath}${path}`
+
+  const unixTime = await getTime();
+  const body = { amount };
+  const jsonBody = JSON.stringify(body);
+  const signedPayload = generateSignature(unixTime, method, path, jsonBody, apiSecret);
+
+  const response = await fetch(url, 
+    method === "POST"? {
+        method,
+        headers: {
+        "Content-Type": "application/json",
+        "YAYA-API-KEY": apiKey,
+        "YAYA-API-TIMESTAMP": unixTime,
+        "YAYA-API-SIGN": signedPayload,
+        },
+        body: jsonBody,
+    }:
+    {
+        method,
+        headers: {
+        "Content-Type": "application/json",
+        "YAYA-API-KEY": apiKey,
+        "YAYA-API-TIMESTAMP": unixTime,
+        "YAYA-API-SIGN": signedPayload,
+        },
+    });
+
+  const apiResponse = await response.json();
+  if (response.status !== 200 || response.status !== 201) {
+    throw new Error(JSON.stringify({ ...apiResponse }));
+  }
+}
+
+module.exports = apiRequest;
