+++ lib/generateSignature.js
@@ -1,0 +1,10 @@
+const crypto = require('crypto');
+
+function generateSignature(timestamp, method, endpoint, body, apiSecret) {
+    const preHashString = `${timestamp}${method}${endpoint}${body}`;
+    const hmac = crypto.createHmac("sha256", apiSecret);
+    hmac.update(preHashString);
+    return Buffer.from(hmac.digest()).toString("base64");
+}
+
+module.exports = generateSignature;
