+++ utils/constants.js
@@ -1,5 +1,7 @@
 const CONSTANTS = {
   TimestampExpiry: 300,
+  baseUrl: "https://yayawallet.com/api/en",
+  apiPath: "/api/en",
 }
 
 module.exports = CONSTANTS;
